pyproject.toml
[tool.poetry]
name = "swarm-example"
version = "0.1.0"
description = ""
authors = ["Connor (Cub) <connor@cubdigital.co.nz>"]
readme = "README.md"
packages = [{ include = "swarm_example" }]

[tool.poetry.dependencies]
python = "^3.12"
swarm = { git = "https://github.com/openai/swarm.git" }
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
test = "swarm_example.run:run"
